Added 3D map objects and got placement/sizing almost right

+++ Ambermoon.Core/Render/ISurface3D.cs
@@ -86,7 +86,7 @@ namespace Ambermoon.Render
 
     public interface ISurface3DFactory
     {
-        ISurface3D Create(SurfaceType type, int width, int height, uint textureWidth, uint textureHeight,
+        ISurface3D Create(SurfaceType type, float width, float height, uint textureWidth, uint textureHeight,
             uint mappedTextureWidth, uint mappedTextureHeight, WallOrientation wallOrientation = WallOrientation.Normal,
             int textureAtlasX = 0, int textureAtlasY = 0);
     }
